Revert "🔥 글자인식: 명령어 제거"
