+++ src/index.ts
@@ -11,7 +11,6 @@ declare module "discord.js" {
     commands: Collection<string, any>;
     buttons: Collection<string, any>;
     error: typeof import("./handler/error.js").default;
-    achievements: typeof import("./handler/achievements.js").default;
   }
 }
 
@@ -24,7 +23,6 @@ const client = new Client({
 client.commands = new Collection();
 client.buttons = new Collection();
 client.error = (await import("./handler/error.js")).default;
-client.achievements = (await import("./handler/achievements.js")).default;
 const player = Player.singleton(client, {
   ytdlOptions: {
     quality: "highestaudio",
